pyproject.toml
[project]
name = "ytmigratewl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
    "yt-dlp>=2025.9.5",
]
